Move EEID attester plugin and install headers
